05-monitor-and-restart-vps/requirements.txt
attrs==23.1.0
bcrypt==4.2.0
boto3==1.35.31
botocore==1.35.31
cattrs==23.1.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Deprecated==1.2.14
docstring-to-markdown==0.12
et-xmlfile==1.1.0
exceptiongroup==1.1.3
idna==3.10
importlib-metadata==6.8.0
jedi==0.19.1
jedi-language-server==0.41.1
jmespath==1.0.1
linode_api4==5.22.0
lsprotocol==2023.0.0b1
microvenv==2023.5.post1
openpyxl==3.1.5
packaging==24.1
paramiko==3.5.0
parso==0.8.3
polling==0.3.2
pycparser==2.22
pygls==1.0.2
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.2
schedule==1.2.2
six==1.16.0
tomli==2.0.1
typeguard==3.0.2
typing_extensions==4.8.0
urllib3==2.2.3
wrapt==1.16.0
zipp==3.19.1
